generator/requirements.txt
PyQt5
pillow
requests

# # For Single exe 
pyinstaller ; extras ~= "build"
